+++ visualization/app/codeCharta/state/angular-redux/effects/effects.module.ts
@@ -1,4 +1,4 @@
-import { APP_INITIALIZER, Inject, InjectionToken, ModuleWithProviders, NgModule, Optional, SkipSelf } from "@angular/core"
+import { APP_INITIALIZER, InjectionToken, ModuleWithProviders, NgModule } from "@angular/core"
 import { Action } from "redux"
 import { Observable, Subject } from "rxjs"
 
@@ -11,12 +11,6 @@ export const ActionsToken = new InjectionToken<Actions>("Actions")
 @NgModule({})
 export class EffectsModule {
 	static actions$ = new Subject<Action>()
-
-	constructor(@Optional() @SkipSelf() @Inject(EffectsModule) parentModule?: EffectsModule) {
-		if (parentModule) {
-			throw new Error("EffectsModule is already loaded. Import it in the AppModule only")
-		}
-	}
 
 	static forRoot(effects = []): ModuleWithProviders<EffectsModule> {
 		return {