+++ visualization/app/codeCharta/ui/metricChooser/metricChooser.component.ts
@@ -1,21 +1,31 @@
 import "./metricChooser.component.scss"
-import { Component, ElementRef, Inject, Input, ViewChild } from "@angular/core"
+import { Component, ElementRef, Inject, Input, OnInit, ViewChild } from "@angular/core"
 import { Store } from "../../state/angular-redux/store"
 import { nodeMetricDataSelector } from "../../state/selectors/accumulatedData/metricData/nodeMetricData.selector"
+import { Observable } from "rxjs"
+import { EdgeMetricData, NodeMetricData } from "../../codeCharta.model"
+import { edgeMetricDataSelector } from "../../state/selectors/accumulatedData/metricData/edgeMetricData.selector"
+
+type MetricChooserType = "node" | "edge"
 
 @Component({
 	selector: "cc-metric-chooser",
 	template: require("./metricChooser.component.html")
 })
-export class MetricChooserComponent {
+export class MetricChooserComponent implements OnInit {
 	@Input() selectedMetricName: string
 	@Input() searchPlaceholder: string
 	@Input() handleMetricChanged: (newSelectedMetricName: string) => void
+	@Input() type: MetricChooserType = "node"
 	@ViewChild("searchTermInput") searchTermInput: ElementRef<HTMLInputElement>
-	nodeMetricData$ = this.store.select(nodeMetricDataSelector)
 	searchTerm = ""
+	metricData$: Observable<NodeMetricData[] | EdgeMetricData[]>
 
 	constructor(@Inject(Store) private store: Store) {}
+
+	ngOnInit(): void {
+		this.metricData$ = this.store.select(this.type === "node" ? nodeMetricDataSelector : edgeMetricDataSelector)
+	}
 
 	handleOpenedChanged(opened: boolean) {
 		if (opened) {