+++ visualization/app/codeCharta/ui/metricChooser/metricChooser.module.ts
@@ -2,12 +2,12 @@ import { CommonModule } from "@angular/common"
 import { NgModule } from "@angular/core"
 import { FormsModule } from "@angular/forms"
 import { MaterialModule } from "../../../material/material.module"
-import { FilterNodeMetricDataBySearchTermPipe } from "./filterNodeMetricDataBySearchTerm.pipe"
+import { FilterMetricDataBySearchTermPipe } from "./filterMetricDataBySearchTerm.pipe"
 import { MetricChooserComponent } from "./metricChooser.component"
 
 @NgModule({
 	imports: [CommonModule, MaterialModule, FormsModule],
-	declarations: [MetricChooserComponent, FilterNodeMetricDataBySearchTermPipe],
+	declarations: [MetricChooserComponent, FilterMetricDataBySearchTermPipe],
 	exports: [MetricChooserComponent],
 	entryComponents: [MetricChooserComponent]
 })