+++ visualization/app/codeCharta/ui/metricChooser/metricChooserTypeHovered/metricChooserTypeHovered.module.ts
@@ -1,0 +1,10 @@
+import { CommonModule } from "@angular/common"
+import { NgModule } from "@angular/core"
+import { MetricChooserTypeHoveredComponent } from "./metricChooserTypeHovered.component"
+
+@NgModule({
+	imports: [CommonModule],
+	declarations: [MetricChooserTypeHoveredComponent],
+	exports: [MetricChooserTypeHoveredComponent]
+})
+export class MetricChooserTypeHoveredModule {}